adding rule to avoid d! messages being deleted

+++ bot.py
@@ -25,7 +25,7 @@ async def on_ready():
 @bot.event
 async def on_message(message):
     channel_role = channelroledao.get(message.channel.id)
-    if channel_role != None and channel_role.role == 'on-rol':
+    if channel_role != None and channel_role.role == 'on-rol' and 'd!' not in message.content:
         await message.delete()
 
 @bot.command()
